hack to run in meteor 1.3 in web browser
It seems some node-isms are in the tiff.min.js code and are somehow
incompatible with the meteor 1.3 modules implementation. So essentially
I think there's a long-standing bug in tiff.min.js in that "fs" and "path"
modules have require() called on them, but for some reason this wasn't
running in the browser with older versions of Meteor.

A better fix would be welcomed. This will likely cause tiff.min.js to
fail on the server side now.